Update imports in line w/ esModuleInterop refactor
See 8752daeab5cc28a2f82983aa0619255411f45cc8
