+++ src/Environment.tsx
@@ -3,7 +3,7 @@ import { useLoader, useThree } from 'react-three-fiber'
 import { CubeTextureLoader, Texture, PMREMGenerator } from 'three'
 import { RGBELoader } from 'three/examples/jsm/loaders/RGBELoader'
 import { useAsset } from 'use-asset'
-import presetsObj from './helpers/cubemap-assets.json'
+import { presetsObj } from './helpers/cubemap-assets'
 
 function getTexture(texture, gen, isCubeMap) {
   if (isCubeMap) {
