--- src/helpers/cubemap-assets.json
@@ -1,12 +1,0 @@
-{
-  "sunset": "venice_sunset_1k.hdr",
-  "dawn": "kiara_1_dawn_1k.hdr",
-  "night": "dikhololo_night_1k.hdr",
-  "warehouse": "empty_warehouse_01_1k.hdr",
-  "forest": "forest_slope_1k.hdr",
-  "apartment": "lebombo_1k.hdr",
-  "studio": "studio_small_03_1k.hdr",
-  "city": "potsdamer_platz_1k.hdr",
-  "park": "rooitou_park_1k.hdr",
-  "lobby": "st_fagans_interior_1k.hdr"
-}
